The objective of the project is to detect hand gestures seen through webcam of a computer. The python script along with OpenCV lib is used to analyze the live-feed video form the webcam, and process the preprogrammed output. Gesture recognition is becoming one of the most concerned topics, because of the ease of use. Processing live feed from camera and processing it to detect programmed output is widely used in modern TV, Laptops, PSP’s, etc. The idea of working on a project based on hand gesture sensing came from gesture-detection software called flutterapp. https://flutterapp.com/.
@@ -0,0 +1,8 @@
Please open the code Code1.py

make sure there is clear background behind the right-top 1/4th area of image. This area is region of interest

This a gesture control code, which gives differnet output for different with finger values. After running the code.
Please start with lifting your fist or a finger. Then, start increasing the no. and see the output.